Chore/tests (#1)
* chore: test submission_dir

* fix: use absolute path in some paths

* chore: move files out and add test workflow

* fix: test workflow
