+++ .github/workflows/dockerpublish.yml
@@ -1,0 +1,39 @@
+name: Docker
+
+on:
+  push:
+    # Publish `master` as Docker `latest` image.
+    branches:
+      - master
+
+    # Publish `v1.2.3` tags as releases.
+    tags:
+      - v*
+
+  # Run tests for any PRs.
+  pull_request:
+
+env:
+  # TODO: Change variable to your image's name.
+  IMAGE_NAME: judger-test
+  PROJECT_ROOT: /usr/src/judger
+
+jobs:
+  # Run tests.
+  # See also https://docs.docker.com/docker-hub/builds/automated-testing/
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Run tests
+        run: |
+          if [ -f docker-compose.test.yml ]; then
+            docker-compose --file docker-compose.test.yml build
+            docker-compose --file docker-compose.test.yml run sut
+          else
+            docker build . --file Dockerfile --tag $IMAGE_NAME
+            docker run --name test --entrypoint=python -w $PROJECT_ROOT/executors/tests \
+                       -e PYTHONPATH=$PROJECT_ROOT --rm $IMAGE_NAME -m unittest -v
+          fi
