+++ executors/tests/test_gxx.py
@@ -4,14 +4,14 @@ from executors import get_executor
 from consts import VerdictResult
 import exceptions
 import shutil
+import conf
 
 
 class TestGxx(unittest.TestCase):
     submission_lang = "GXX"
-    submission_dir = f'problems/1/test'
+    submission_dir = f'{conf.PROJECT_ROOT}/executors/tests/problems/1/test'
 
     def test_AC(self):
-        self.submission_dir = f'problems/1/test'
         ac_code = """
         #include <iostream>
         using namespace std;
@@ -34,7 +34,6 @@ class TestGxx(unittest.TestCase):
         self.assertEqual(result['verdict'], VerdictResult.AC)
 
     def test_WA(self):
-        self.submission_dir = f'problems/1/test'
         wa_code = """
             #include <iostream>
             using namespace std;
@@ -58,7 +57,6 @@ class TestGxx(unittest.TestCase):
         self.assertEqual(result['verdict'], VerdictResult.WA)
 
     def test_TLE(self):
-        self.submission_dir = f'problems/1/test'
         tle_code = """
             #include <iostream>
             using namespace std;
@@ -85,7 +83,6 @@ class TestGxx(unittest.TestCase):
         self.assertEqual(result["verdict"], VerdictResult.TLE)
 
     def test_MLE(self):
-        self.submission_dir = f'problems/1/test'
         mle_code = """
             #include <iostream>
             int fuck[3000000];
@@ -114,7 +111,6 @@ class TestGxx(unittest.TestCase):
         self.assertEqual(result["verdict"], VerdictResult.MLE)
 
     def test_RE(self):
-        self.submission_dir = f'problems/1/test'
         re_code = """
             #include <iostream>
             int fuck[60000];
@@ -142,7 +138,6 @@ class TestGxx(unittest.TestCase):
         self.assertEqual(result["verdict"], VerdictResult.RE)
 
     def test_CE(self):
-        self.submission_dir = f'problems/1/test'
         ce_code = """
             #include <iostream>
             int fuck[60000000000];
@@ -162,7 +157,6 @@ class TestGxx(unittest.TestCase):
             print(cm.exception)
 
     def test_OLE(self):
-        self.submission_dir = f'problems/1/test'
         ole_code = f"""
                     #include <stdio.h>
                     int main() {{
